clean up CodeSig.compute
